# docs/website/references/http/openapi.json
{
    "components": {
...
        "responses": {
            "error": {
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/error"
                        }
                    }
                },
                "description": "error"
            },
            "success": {
                "description": "ok"
            }
        },
        "schemas": {
...
            "collection_delete": {
                "properties": {
                    "filter": {}
                },
                "type": "object"
            },
            "collection_update": {
                "properties": {
                    "filter": {},
                    "updater": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "delete_result": {
                "properties": {
                    "Count": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "DocIDs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            },
            "document": {
                "additionalProperties": true,
                "type": "object"
            },
            "encrypted_index": {
                "properties": {
                    "FieldName": {
                        "type": "string"
                    },
                    "Type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "encrypted_index_create": {
                "properties": {
                    "FieldName": {
                        "type": "string"
                    },
                    "Type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "error": {
                "properties": {
                    "error": {}
                },
                "type": "object"
            },
...
            "update_result": {
                "properties": {
                    "Count": {
                        "format": "int64",
                        "type": "integer"
                    },
                    "DocIDs": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerToken": {
                "bearerFormat": "JWT",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
...
    "info": {
        "title": "DefraDB API",
        "version": "0"
    },
    "openapi": "3.0.3",
    "paths": {
...
        "/collections/{name}": {
            "delete": {
                "description": "Delete document(s) from a collection",
                "operationId": "collection_delete_with_filter",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/collection_delete"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/delete_result"
                                }
                            }
                        },
                        "description": "Delete results"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "collection"
                ]
            },
            "get": {
                "description": "Get all document IDs",
                "operationId": "collection_keys",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "collection"
                ]
            },
            "patch": {
                "description": "Update document(s) in a collection",
                "operationId": "collection_update_with_filter",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/collection_update"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/update_result"
                                }
                            }
                        },
                        "description": "Update results"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "collection"
                ]
            },
            "post": {
                "description": "Create document(s) in a collection",
                "operationId": "collection_create",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "oneOf": [
                                    {
                                        "$ref": "#/components/schemas/document"
                                    },
                                    {
                                        "items": {
                                            "$ref": "#/components/schemas/document"
                                        },
                                        "type": "array"
                                    }
                                ]
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "collection"
                ]
            }
        },
        "/collections/{name}/encrypted-indexes": {
            "get": {
                "description": "List encrypted indexes",
                "operationId": "encrypted_index_list",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/encrypted_index"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "List of encrypted indexes"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "encrypted_index"
                ]
            },
            "post": {
                "description": "Create an encrypted index",
                "operationId": "encrypted_index_create",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/encrypted_index_create"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/encrypted_index"
                                }
                            }
                        },
                        "description": "Encrypted index description"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "encrypted_index"
                ]
            }
        },
        "/collections/{name}/encrypted-indexes/{field}": {
            "delete": {
                "description": "Delete an encrypted index",
                "operationId": "encrypted_index_delete",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Field name of the encrypted index",
                        "in": "path",
                        "name": "field",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "encrypted_index"
                ]
            }
        },
...
        "/collections/{name}/indexes/{index}": {
            "delete": {
                "description": "Delete a secondary index",
                "operationId": "index_drop",
                "parameters": [
                    {
                        "description": "Collection name",
                        "in": "path",
                        "name": "name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "index",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/components/responses/success"
                    },
                    "400": {
                        "$ref": "#/components/responses/error"
                    },
                    "default": {
                        "description": ""
                    }
                },
                "tags": [
                    "index"
                ]
            }
        },
...
    },
    "servers": [
        {
            "description": "Local DefraDB instance",
            "url": "/api/v0"
        }
    ],
...
}